requests pulls in urllib3 2.x but that fails with an error
They say 1.26.x will be maintained for a while.  To fix, someone needs
to figure out what the exact problem is as shown in:
https://github.com/psf/requests/issues/6443#issuecomment-1568210854

I suppose there will need to be some change related to the urllib 2.x
migration, but it is unclear to me which.  The usage of requests in
fdroid/sdkmanager is quite simple.
https://urllib3.readthedocs.io/en/latest/v2-migration-guide.html